updates way to use the custom variants
