+++ src/types/index.d.ts
@@ -10,3 +10,4 @@ export * from './graph';
 export * from './material-ui';
 export * from './recoil';
 export * from './selfid';
+export * from './ape-text-field';
